initial webui commit
- You can get plain JavaScript object representing the state of the
interface by calling viewModel.toPlainObject(). Note that this only
includes the relevant parts of the state for building a query. This is
what the debug section at the bottom of the screen shows.
- To populate the suggestions bar along the left side, you can do a
call like viewModel.suggestions(["foo","bar"]). This will update the
interface automatically.
- Suggestions are drag and drop enabled, although only for creating new
clauses / buckets. I'm figuring out if there is an efficient way to
have the query drag and drop stuff added to DOM nodes that are created
after the page has loaded.
- To turn off debug mode (for now, this means getting rid of the JSON
viewer along the bottom), set viewModel.debug = false before
the ko.applyBindings(viewModel) line.
- Most of the interface will self-update if you programmatically modify
members of viewModel using the auto-generated setters, but there is one
important exception: the year slider. Of course, it works fine in the
other direction, so dragging the slider will modify viewModel as
expected. Although this isn't a big issue, it bothers me enough that I
will spend some more time looking for a fix.